requirements.txt
alembic==1.8.1
anyio==3.6.1
asyncpg==0.26.0
bcrypt==4.0.0
cffi==1.15.1
click==8.1.3
cryptography==37.0.4
dnspython==2.2.1
ecdsa==0.18.0
email-validator==1.2.1
fastapi==0.81.0
greenlet==1.1.3
h11==0.13.0
idna==3.3
importlib-metadata==4.12.0
importlib-resources==5.9.0
Mako==1.2.2
MarkupSafe==2.1.1
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.2
python-dotenv==0.20.0
python-jose==3.3.0
python-multipart==0.0.5
rsa==4.9
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.40
starlette==0.19.1
typing==3.7.4.3
typing-extensions==4.3.0
uvicorn==0.18.3
zipp==3.8.1
pre-commit==2.20.0
